pyproject.toml
[project]
name = "ai-app-py-only"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepeval>=3.7.0",
    "langchain>=1.0.7",
    "langchain-openai>=1.0.3",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
